+++ hdf/test/nbit.c
@@ -289,7 +289,7 @@ test_nbit3(int32 fid)
     CHECK_VOID(ret, FAIL, "Hendaccess");
 
     MESSAGE(5, printf("Verifying data\n"););
-    memset(convbuf, 0, DFKNTsize(DFNT_UINT16) * NBIT_SIZE3);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_UINT16) * NBIT_SIZE3));
     ret = Hgetelement(fid, NBIT_TAG3, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE3 * DFKNTsize(DFNT_UINT16)) {
         HEprint(stderr, 0);
@@ -366,7 +366,7 @@ test_nbit4(int32 fid)
 
     MESSAGE(5, printf("Verifying data\n"););
 
-    memset(convbuf, 0, DFKNTsize(DFNT_INT16) * NBIT_SIZE4);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_INT16) * NBIT_SIZE4));
 
     ret = Hgetelement(fid, NBIT_TAG4, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE4 * DFKNTsize(DFNT_INT16)) {
@@ -444,7 +444,7 @@ test_nbit5(int32 fid)
 
     MESSAGE(5, printf("Verifying data\n"););
 
-    memset(convbuf, 0, DFKNTsize(DFNT_UINT32) * NBIT_SIZE5);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_UINT32) * NBIT_SIZE5));
 
     ret = Hgetelement(fid, NBIT_TAG5, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE5 * DFKNTsize(DFNT_UINT32)) {
@@ -523,7 +523,7 @@ test_nbit6(int32 fid)
 
     MESSAGE(5, printf("Verifying data\n"););
 
-    memset(convbuf, 0, DFKNTsize(DFNT_INT32) * NBIT_SIZE6);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_INT32) * NBIT_SIZE6));
 
     ret = Hgetelement(fid, NBIT_TAG6, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE6 * DFKNTsize(DFNT_INT32)) {
@@ -725,7 +725,7 @@ test_nbit9(int32 fid)
 
     MESSAGE(5, printf("Verifying data\n"););
 
-    memset(convbuf, 0, DFKNTsize(DFNT_UINT16) * NBIT_SIZE9);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_UINT16) * NBIT_SIZE9));
 
     ret = Hgetelement(fid, NBIT_TAG9, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE9 * DFKNTsize(DFNT_UINT16)) {
@@ -804,7 +804,7 @@ test_nbit10(int32 fid)
 
     MESSAGE(5, printf("Verifying data\n"););
 
-    memset(convbuf, 0, DFKNTsize(DFNT_INT16) * NBIT_SIZE10);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_INT16) * NBIT_SIZE10));
 
     ret = Hgetelement(fid, NBIT_TAG10, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE10 * DFKNTsize(DFNT_INT16)) {
@@ -898,7 +898,7 @@ test_nbit11(int32 fid)
 
     MESSAGE(5, printf("Verifying data\n"););
 
-    memset(convbuf, 0, DFKNTsize(DFNT_UINT32) * NBIT_SIZE11);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_UINT32) * NBIT_SIZE11));
 
     ret = Hgetelement(fid, NBIT_TAG11, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE11 * DFKNTsize(DFNT_UINT32)) {
@@ -978,7 +978,7 @@ test_nbit12(int32 fid)
 
     MESSAGE(5, printf("Verifying data\n"););
 
-    memset(convbuf, 0, DFKNTsize(DFNT_INT32) * NBIT_SIZE12);
+    memset(convbuf, 0, (size_t)(DFKNTsize(DFNT_INT32) * NBIT_SIZE12));
 
     ret = Hgetelement(fid, NBIT_TAG12, (uint16)ref1, convbuf);
     if (ret != NBIT_SIZE12 * DFKNTsize(DFNT_INT32)) {
