+++ hdf/util/hdf2gif.c
@@ -169,7 +169,7 @@ main(int argc, char **argv)
         }
         /* End BUG FIX 601 */
 
-        Image  = (BYTE *)malloc(dim_sizes[0] * dim_sizes[1]);
+        Image  = (BYTE *)malloc((size_t)(dim_sizes[0] * dim_sizes[1]));
         status = GRreadimage(ri_id, start, stride, dim_sizes, Image);
 
         /* If the first image does not have a palette, I make my own global color table
